ml-agent/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# OpenAI
openai>=1.50.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Date parsing
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0

# Type checking
mypy>=1.8.0

# Logging
structlog>=24.0.0
# Database ORM
prisma>=0.15.0  # Prisma Client Python for PostgreSQL

# Web server
fastapi>=0.115.0
uvicorn>=0.30.0
